requirements.txt
livekit-agents>=0.12.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-elevenlabs>=0.7.0
livekit-plugins-openai>=0.10.0
livekit-plugins-silero>=0.7.0
python-dotenv>=1.0.1
langdetect>=1.0.9
pydantic>=2.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
